pyproject.toml
[tool.poetry]
name = "OpenSlots-lambda"
version = "0.1.0"
description = "Appointment booking platform with AWS Lambda"
authors = ["Dev Team"]

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.28"
stripe = "^7.0"
pydantic = "^2.0"
PyJWT = "^2.8"

[tool.poetry.group.infra.dependencies]
aws-cdk-lib = "^2.100"
constructs = "^10.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4"
pytest-cov = "^4.1"
black = "^23.0"
mypy = "^1.5"
ruff = "^0.1"
moto = "^5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
